MNT Refactor and reorganize code to help prevent future circular imports
