+++ lute/tasks/task.py
@@ -3,11 +3,6 @@
 Classes:
     Task: Abstract base class from which all analysis tasks are derived.
 
-    TaskResult: Output of a specific analysis task.
-
-    TaskStatus: Enumeration of possible Task statuses (running, pending, failed,
-        etc.).
-
     BinaryTask: Class to run a third-party executable binary as a `Task`.
 """
 
@@ -16,21 +11,20 @@ __author__ = "Gabriel Dorlhiac"
 
 import time
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
 from typing import Any, List, Dict, Union, Type, TextIO, Optional
-from enum import Enum
 import os
 import warnings
 import signal
 import types
 
-from ..io.config import (
+from ..io.models.base import (
     TaskParameters,
     ThirdPartyParameters,
     TemplateConfig,
     AnalysisHeader,
 )
 from ..execution.ipc import *
+from .dataclasses import *
 
 if __debug__:
     warnings.simplefilter("default")
@@ -55,75 +49,6 @@ if __debug__:
     warnings.showwarning = lute_warn
 else:
     warnings.simplefilter("ignore")
-
-
-class TaskStatus(Enum):
-    """Possible Task statuses."""
-
-    PENDING = 0
-    """
-    Task has yet to run. Is Queued, or waiting for prior tasks.
-    """
-    RUNNING = 1
-    """
-    Task is in the process of execution.
-    """
-    COMPLETED = 2
-    """
-    Task has completed without fatal errors.
-    """
-    FAILED = 3
-    """
-    Task encountered a fatal error.
-    """
-    STOPPED = 4
-    """
-    Task was, potentially temporarily, stopped/suspended.
-    """
-    CANCELLED = 5
-    """
-    Task was cancelled prior to completion or failure.
-    """
-    TIMEDOUT = 6
-    """
-    Task did not reach completion due to timeout.
-    """
-
-
-@dataclass
-class TaskResult:
-    """Class for storing the result of a Task's execution with metadata.
-
-    Attributes:
-        task_name (str): Name of the associated task which produced it.
-
-        task_status (TaskStatus): Status of associated task.
-
-        summary (str): Short message/summary associated with the result.
-
-        payload (Any): Actual result. May be data in any format.
-
-        impl_schemas (str): A string listing `Task` schemas implemented by the
-            associated `Task`. Schemas define the category and expected output
-            of the `Task`. An individual task may implement/conform to multiple
-            schemas. Multiple schemas are separated by ';', e.g.
-                * impl_schemas = "schema1;schema2"
-    """
-
-    task_name: str
-    task_status: TaskStatus
-    summary: str
-    payload: Any
-    impl_schemas: Optional[str] = None
-
-
-@dataclass
-class DescribedAnalysis:
-    task_result: TaskResult
-    task_parameters: TaskParameters
-    task_env: Dict[str, str]
-    poll_interval: float
-    communicator_desc: List[str]
 
 
 class Task(ABC):
